decoding of GNSS position information added
